Fixed the XYZ calibration for all variants.
Thanks for the wonderful work and making all the variants explicit, this was really helpful!
I had a concern with respect to Bondtech BMG and MK3S firmware, and the way Prusa fixed it and how for a bondtech it reduces Z_MAX_POS.

Basically, instead of increasing Z_MAX_POS for the MK3S by 6, Prusa decided to leave it as is,
and... to extend the travel down when calibrating XYZ by 7.
As a result, Prusa's "S" firmware does not work for a taller extruder, like the Bondtech for the "S" variant.

What Prusa _should_ have done is INCREASE Z_MAX_POS in their variant by 6 for the "S" extruder and
leave the "travel down mechanism"for XYZ calibration as it was.
Then we could have kept the Z_MAX_POS to 210 for BMG variants and all would be fine.

Now, we have a smaller Z_MAX_POS as necessary, if we want to use the "S" firmware.

I implemented it - in my opinion - correctly in the branch, by introducing a new variable to work exactly "where it hurts":
during Z calibration: see the pull request.

In particular, look at Marlin_main.cpp at line 2121 and the "S" variants, that now set the offset correctly (basically to 9 for "S" extruders and 2 elsewhere..)
There is also an unclear thing in the Prusa file, in that they test for "MK3", (the alternative being MK3S), but the should have tested on MK3S and MK2.5S and then have all the others as alternative.
My variant puts that test "out of order" and sets the desired XYZ calibration travel to the correct value for each extruder.

I issued a pull request to Prusa, but have little hope that they would do it correctly, or even better: set Z_MAX_POS for their "S" variants to 217 and leave the algorithm in Marlin_main.cpp like how it was before, then Prusa owners would "enjoy" higher Z....

To summarize: the change is easy: only one file (marlin_main.cpp) and a new constant in each variant. I just went through the many variant files to set it.

I hope it is clear (and I believe it to be cleaner).